Fix 8 and half of 7

+++ app/controllers/phones_controller.rb
@@ -3,7 +3,6 @@ class PhonesController < ApplicationController
 
   def create
     current_participant.phones.create! params[:phone]
-    
     render json: { _success: true, _alert: 'added', _path_to_go: '' }
   rescue StandardError => e
     render json: { _success: false, _alert: 'error' }
@@ -11,6 +10,7 @@ class PhonesController < ApplicationController
 
   def destroy
     phone = current_participant.phones.find params[:id]
+    return render json: { _success: false, _alert: 'cannot' } if Claim.where(phone_id: phone.id).empty?
     phone.destroy
     render json: { _success: true, _alert: 'deleted' }
   end
